updated the way models were parsed
This update removed global variables that were being used in those functions, in favor of passing in the model that is being manipulated.

Also removes the logic that was being used to resolve dependencies in favor of using `dependency-resolver` that already handles this functionality. This will reduce the amount of testing that we have to do around this file.