Feature/1.0.0/main structure (#103)
* added the ability to add documentation

This will allow us to document the api to make it easier to know what's avaiable and what's actually going on.

* removed Yarn for now

This is because yarn doesn't handle binaries correctly, and also doesn't handle things like couchbase which have node-gyp and that have to compile when it's being installed.

* Moved the cli to to be under `cli.js`

* Renamed `generator.js` to be `index.js`

This is so that the main file will  be the `index.js` file. This way we can export it and get the same functionality.

* added the Logger to the utils

This will be the base logger that is used throughout the application

* added documentation to the other util functions

* added support for ci testing

* added unit tests and bug fixes for `utils.js`

* syntax update

* added `base.js` which will be the base for all the classes

This will allow you to run each of the functionalities separately without having to depend on each other. It will also allow the logging to be consistent throught the app.

* added ability to pass a glob to `utils.findFiles`

* updated the way models were parsed

This update removed global variables that were being used in those functions, in favor of passing in the model that is being manipulated.

Also removes the logic that was being used to resolve dependencies in favor of using `dependency-resolver` that already handles this functionality. This will reduce the amount of testing that we have to do around this file.

* added babel-external-helpers

* Updated input to be a class

* updated models to be a class

Also drastically simplified the exisitng code by reworking a few functions and using dependency-resolver to resolve the different dependencies for each model

* updated the main fakit function to be a class

This main function extends Inputs, Models, and the Base

* Updated the cli to only pass valid options.

* Updated documents to be a class

Also updated to generate documents in documents instead of in models

* Updated to return models as an array

This update is to remove the need for `this.model_order` by ordering them to begin with.

* Cleaned up the cli

Also added several todo items to help the various changes that need to take place before the release.

* replaced `--format` with `--spacing`

* Added the base for the new Output class

This also adds tests for validation and options that have been passed.

* Added the base functionality to the outputters

* Finished folder output

Complete with tests

* added a pool utility function

This is a function that is similar to `map` but it will limit the concurrancy of the functions that're running.

* updated archive option to be a filename

* added the zip output type

complete with tests

* general cleanup

* removed the need for output document

* added the `key` generated document

* updated the way the setup function determins the outputter

This update is to accomidate the `return` option

* update the file generation for the zip test

This just updates the location as to where the files get created.

* added the base for `Output.output`

This also adds test cases for the different types and languages.

* added a few prepare tests for output

* added the finalize function

* added the timeout option that's used for servers

* setup the rest of the sync-gateway output

* setup the rest of couchbase output

* removed `object-path` in favor of lodash `set` and `get`

* cleaned up dependencies

* fixed bug with the verbose logging option

* Updated couchbase output to use `couchbase-promises`

Also added test cases for them to ensure it's working correctly

* Added a few tests for sync-gateway

* remove the Documents class because it's no longer needed

* Updated the main fakit class

* added nyc code coverage

* added the basis for testing the data generation

* 🗑removed the `output.js`