Issue #136: Store documents in Couchbase as JSON data format
Motivation
----------
Currently, when outputting to Couchbase documents are marked with the
string data format.  This can cause compatibility issues with some
Couchbase SDKs, such as .Net.

Modifications
-------------
When outputting to Couchbase with the JSON format, use an output parser
that does nothing and returns in the in-memory object.

Results
-------
JSON is serialized by the Couchbase NodeJS SDK, so it sets the data
format correctly as JSON instead of string.